refactor(ui): separate LevelButton component
